requirements.txt
fastapi
google-generativeai
requests
uvicorn
